requirements.txt
pandas>=2.1,<3.0
pymongo>=4.6
sqlalchemy>=2.0,<3.0
psycopg2-binary>=2.9
python-dotenv>=1.0
openpyxl>=3.1
